requirements.txt
numpy
Pillow==3.2.0
tensorflow==0.11.0
